improved conversion to Poly and made its own function
